moves field to its own file.
